Move graphite support functions to seperate module
